Fix resource issue
